chore: move the network vat to the vats package
