--- packages/SwingSet/src/vats/network/index.js
@@ -1,4 +1,0 @@
-export * from './network.js';
-export { default as makeRouter } from './router.js';
-export * from './multiaddr.js';
-export * from './bytes.js';
