+++ packages/pegasus/src/ibc-trace.js
@@ -2,7 +2,7 @@
 import { Far } from '@endo/far';
 import { assert, details as X } from '@agoric/assert';
 
-import { parse } from '@agoric/swingset-vat/src/vats/network/multiaddr.js';
+import { parse } from '@agoric/vats/src/network/multiaddr.js';
 
 /**
  * Return a source-prefixed version of the denomination, as specified in
