+++ packages/pegasus/test/test-peg.js
@@ -5,7 +5,7 @@ import { E, Far } from '@endo/far';
 import {
   makeNetworkProtocol,
   makeLoopbackProtocolHandler,
-} from '@agoric/swingset-vat/src/vats/network/index.js';
+} from '@agoric/vats/src/network/network.js';
 
 import bundleSource from '@endo/bundle-source';
 import { AmountMath } from '@agoric/ertp';
@@ -85,7 +85,7 @@ async function testRemotePeg(t) {
   /**
    * Pretend we're Gaia.
    *
-   * @type {import('@agoric/swingset-vat/src/vats/network').Connection?}
+   * @type {import('@agoric/vats/src/network').Connection?}
    */
   let gaiaConnection;
   E(portP).addListener(
