+++ packages/vats/src/network/bytes.js
