+++ packages/vats/src/network/vat-network.js
@@ -1,5 +1,5 @@
 // @ts-check
-import { makeRouterProtocol } from '@agoric/swingset-vat/src/vats/network/router.js';
+import { makeRouterProtocol } from './router.js';
 
 export function buildRootObject(_vatPowers) {
   return makeRouterProtocol(); // already Far('Router')
